feat(core+cli): share tool invocation/result formatting between desktop and TUI
Extract formatToolInvocationLine, formatQuietJsonValue, extractToolCommand,
and formatAsKeyValueLines from packages/ui/src/tool-activity/builtin-preview.ts
into a new pure module in @maka/core (tool-quiet-preview.ts), parameterizing
the 7 hardcoded Chinese strings as a locale catalog (zh/en).

- @maka/ui re-imports from @maka/core; desktop passes detectUiLocale() —
  output is byte-identical to the previous inline implementation.
- The TUI (pi-transcript-tools.ts) replaces the JSON.stringify fallbacks in
  toolInputSummary and plainResultText with the shared formatters, passing
  'en'. AskUserQuestion, GoalSet, Automation, Skill, and any future tool
  without a custom case render human-readable text instead of raw JSON.
- This also fixes Bug 1: Skill tool args {"name":"x"} no longer dump as
  raw JSON — the HEADLINE_KEYS path extracts the skill name.

Closes #1065.
